H1_high_level_mujoco/requirements.txt
torch
torchvision
torchaudio
mediapipe
opencv-python
ultralytics
mistralai
openai-whisper
pyaudio
pygame
langdetect
gtts
transformers
easyocr
segment_anything
sacremoses
open3d